Refactor Int type to Long type in MC stats library (#1256)
## TL;DR
This pull request changes the data type of `Reach` and `Impression`
variables from Int to Long in multiple functions and test cases within
the code.

## Why make this change
- The change to the data type of the "totalReach" variable from Int to
Long allows for larger values to be used in the computation of
variances.
- Updating the data classes to use Long data types instead of Int for
certain variables ensures consistency and accuracy in the computation of
variances.
- The change improves the precision and accuracy of the computations,
especially when dealing with large reach values.

+++ src/main/kotlin/org/wfanet/measurement/measurementconsumer/stats/Variances.kt
@@ -90,7 +90,7 @@ object Variances {
    * Different types of frequency histograms have different values of [multiplier].
    */
   private fun deterministicFrequencyRelativeVariance(
-    totalReach: Int,
+    totalReach: Long,
     reachRatio: Double,
     measurementParams: FrequencyMeasurementParams,
     multiplier: Int
@@ -114,7 +114,7 @@ object Variances {
    * Reach count = [totalReach] * [reachRatio]
    */
   private fun deterministicFrequencyCountVariance(
-    totalReach: Int,
+    totalReach: Long,
     totalReachVariance: Double,
     reachRatio: Double,
     reachRatioVariance: Double,
@@ -204,7 +204,7 @@ object Variances {
     sketchParams: LiquidLegionsSketchParams,
     measurementParams: FrequencyMeasurementParams,
   ): (
-    totalReach: Int,
+    totalReach: Long,
     reachRatio: Double,
     measurementParams: FrequencyMeasurementParams,
     multiplier: Int
@@ -273,7 +273,7 @@ object Variances {
     sketchParams: LiquidLegionsSketchParams,
     measurementParams: FrequencyMeasurementParams,
   ): (
-    totalReach: Int,
+    totalReach: Long,
     reachRatio: Double,
     measurementParams: FrequencyMeasurementParams,
     multiplier: Int
@@ -337,14 +337,14 @@ object Variances {
     params: FrequencyMeasurementVarianceParams,
     frequencyRelativeVarianceFun:
       (
-        totalReach: Int,
+        totalReach: Long,
         reachRatio: Double,
         measurementParams: FrequencyMeasurementParams,
         multiplier: Int
       ) -> Double,
     frequencyCountVarianceFun:
       (
-        totalReach: Int,
+        totalReach: Long,
         totalReachVariance: Double,
         reachRatio: Double,
         reachRatioVariance: Double,
